+++ src/pages/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { App } from '../App';
+import { App } from '../containers/App';
 
 const Home: React.FC = () => <App></App>;
